% Auto-generated: do not edit by hand
\name{simpleEditor}

\alias{simpleEditor}

\title{SimpleEditor component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
simpleEditor(id=NULL, editorState=NULL, placeholder=NULL)
}

\arguments{
\item{id}{Character. }

\item{editorState}{Named list. }

\item{placeholder}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

